API CHANGE: Make indexes and slice borders signed (int64_t)
For abstract index spaces, it makes no sense to force them to be
positive. This is also bad for calculating neighbors at nearside
indexes which easily generates underflows.

However, local indexes and element counts within slices stay.
The reasoning for local indexes is that they always start from
a given base, covering a range of a given number of elements.

+++ src/partitioner.c
@@ -554,7 +554,7 @@ void runReassignPartitioner(Laik_Partitioner* pr,
         int to   = oldBA->tslice[sliceNo].s.to.i[0];
 
         if (data->getIdxW) {
-            for(uint64_t i = from; i < to; i++) {
+            for(int64_t i = from; i < to; i++) {
                 idx.i[0] = i;
                 totalWeight += (data->getIdxW)(&idx, data->userData);
             }
@@ -587,11 +587,11 @@ void runReassignPartitioner(Laik_Partitioner* pr,
         }
 
         // re-distribute
-        int from = oldBA->tslice[sliceNo].s.from.i[0];
-        int to = oldBA->tslice[sliceNo].s.to.i[0];
+        int64_t from = oldBA->tslice[sliceNo].s.from.i[0];
+        int64_t to = oldBA->tslice[sliceNo].s.to.i[0];
 
         slc.from.i[0] = from;
-        for(uint64_t i = from; i < to; i++) {
+        for(int64_t i = from; i < to; i++) {
             if (data->getIdxW) {
                 idx.i[0] = i;
                 weight += (data->getIdxW)(&idx, data->userData);
